# Refactor implementation of F.compute_deltas