Removing the `repeat` function and replacing groups convolution with regular convolution can lower the size of kernel, and is more memory friendly. 
Can also have some performance gain, probably.