@mthrok 2021-04-05
@yoyololicon 

Could you help benchmark the result?
I am getting twice slower result, so we might need to revert the merge.

<details><summary>script</summary>

```bash
#!/usr/bin/env bash

python -m torch.utils.collect_env

OMP_NUM_THREADS=1 numactl --membind 0 --cpubind 0 python -m timeit -s """
import torch
import torchaudio;

batch_size = 32
num_channels = 1
num_freq = 1025
num_frames = 400

torch.manual_seed(0);
spec = torch.randn(batch_size, num_channels, num_freq, num_frames, 2, dtype=torch.float32, device='cpu');
""" """
torchaudio.functional.compute_deltas(spec)
"""
```

</details>

| batch_size |          Before this PR (e964f93)         |         After this PR (8d2eeb1)         |
|:----------:|:-----------------------------------------:|:---------------------------------------:|
| 2          | 50 loops, best of 5: 4.28 msec per loop | 50 loops, best of 5: 8.5 msec per loop |
| 32         | 2 loops, best of 5: 158 msec per loop   |1 loop, best of 5: 443 msec per loop  |