@yoyololicon 2021-04-06
> 2	50 loops, best of 5: 4.28 msec per loop	50 loops, best of 5: 8.5 msec per loop

```
Collecting environment information...
PyTorch version: 1.9.0.dev20210218
Is debug build: False
CUDA used to build PyTorch: Could not collect
ROCM used to build PyTorch: N/A

OS: Ubuntu 20.04.2 LTS (x86_64)
GCC version: (Ubuntu 9.3.0-17ubuntu1~20.04) 9.3.0
Clang version: Could not collect
CMake version: version 3.18.4

Python version: 3.8 (64-bit runtime)
Is CUDA available: False
CUDA runtime version: Could not collect
GPU models and configuration: GPU 0: GeForce RTX 3070
Nvidia driver version: 460.39
cuDNN version: Could not collect
HIP runtime version: N/A
MIOpen runtime version: N/A

Versions of relevant libraries:
[pip3] numpy==1.19.2
[pip3] torch==1.9.0.dev20210218
[pip3] torchaudio==0.9.0a0+e3820e0
[conda] blas                      1.0                         mkl  
[conda] cpuonly                   1.0                           0    pytorch-nightly
[conda] cudatoolkit               11.0.221             h6bb024c_0  
[conda] mkl                       2020.2                      256  
[conda] mkl-service               2.3.0            py38he904b0f_0  
[conda] mkl_fft                   1.2.1            py38h54f3939_0  
[conda] mkl_random                1.1.1            py38h0573a6f_0  
[conda] numpy                     1.19.2           py38h54aff64_0  
[conda] numpy-base                1.19.2           py38hfa32c7d_0  
[conda] pytorch                   1.9.0.dev20210218     py3.8_cpu_0  [cpuonly]  pytorch-nightly
[conda] torchaudio                0.9.0a0+e3820e0           dev_0    <develop>
```


I run on cpu-only, batch size = 2.

Before:
5 loops, best of 5: 67.7 msec per loop

After:
1 loop, best of 5: 957 msec per loop

Well ... that's pretty awful. Have tried different `num_freq` and `num_frames` but can't see any improvement.
Seems groups convolution does have some advantages here.